+++ src/main/java/com/alladeson/caurisit/models/paylaods/ReglementPayload.java
@@ -18,11 +18,11 @@ public class ReglementPayload {
 	// Type de Paiement
 	private Long typePaiementId;
 	// Montant Reçu
-	private Integer montantRecu;
+	private Long montantRecu;
 	// Montant payé
-	private Integer montantPayer;
+	private Long montantPayer;
 	// Montant Rendu
-	private Integer montantRendu;
+	private Long montantRendu;
 	// Description supplémentaire
 	private String description;
 	
@@ -53,37 +53,37 @@ public class ReglementPayload {
 	/**
 	 * @return the montantRecu
 	 */
-	public Integer getMontantRecu() {
+	public Long getMontantRecu() {
 		return montantRecu;
 	}
 	/**
 	 * @param montantRecu the montantRecu to set
 	 */
-	public void setMontantRecu(Integer montantRecu) {
+	public void setMontantRecu(Long montantRecu) {
 		this.montantRecu = montantRecu;
 	}
 	/**
 	 * @return the montantPayer
 	 */
-	public Integer getMontantPayer() {
+	public Long getMontantPayer() {
 		return montantPayer;
 	}
 	/**
 	 * @param montantPayer the montantPayer to set
 	 */
-	public void setMontantPayer(Integer montantPayer) {
+	public void setMontantPayer(Long montantPayer) {
 		this.montantPayer = montantPayer;
 	}
 	/**
 	 * @return the montantRendu
 	 */
-	public Integer getMontantRendu() {
+	public Long getMontantRendu() {
 		return montantRendu;
 	}
 	/**
 	 * @param montantRendu the montantRendu to set
 	 */
-	public void setMontantRendu(Integer montantRendu) {
+	public void setMontantRendu(Long montantRendu) {
 		this.montantRendu = montantRendu;
 	}
 	/**
